Ensure that seeded Redux-form fields define all subfields
Previously, using the value of fields with subfields to create the Redux
form’s field instance caused elements such as unchecked check boxes to 
disappear from the state, which caused issues in re-rendering forms on 
the server.  To fix this, a map of all known subfields for any fields 
that have them is created, and the potential subfield value is pulled 
from the initial values send to `seedFields`.